+++ src/components/HrefActions.vue
@@ -41,7 +41,7 @@
 import { defineAsyncComponent } from 'vue';
 
 import Description from './Description.vue';
-import Utils, { mapMediaTypes } from '../utils';
+import Utils from '../utils';
 import { size, URI } from 'stac-js/src/utils.js';
 import { mapGetters, mapState } from 'vuex';
 import AssetActions from '../../assetActions.config';
@@ -49,9 +49,10 @@ import LinkActions from '../../linkActions.config';
 import AuthUtils from './auth/utils';
 import { Asset } from 'stac-js';
 import { browserProtocols } from 'stac-js/src/http';
-import { imageMediaTypes, zarrMediaTypes } from 'stac-js/src/mediatypes';
+import { imageMediaTypes, geojsonMediaType, wozMediaTypes, zarrMediaTypes } from 'stac-js/src/mediatypes';
 
 const disableDownloadTypes = [...zarrMediaTypes];
+const mapTypes = imageMediaTypes.concat([geojsonMediaType]).concat(wozMediaTypes);
 
 let i = 0;
 
@@ -92,7 +93,8 @@ export default {
   emits: ['show'],
   data() {
     return {
-      id: i++
+      id: i++,
+      tileUrlTemplate: null
     };
   },
   computed: {
@@ -106,7 +108,7 @@ export default {
       return !this.isLoggedIn && this.auth.length > 0;
     },
     tileRendererType() {
-      if (this.buildTileUrlTemplate && !this.useTileLayerAsFallback) {
+      if (this.tileUrlTemplate && !this.useTileLayerAsFallback) {
         return 'server';
       }
       else {
@@ -123,8 +125,8 @@ export default {
       if (typeof this.data?.type !== 'string') {
         return false;
       }
-      // If the tile renderer is a tile server, we can't really know what it supports so we pass all images
-      else if (this.tileRendererType === 'server' && imageMediaTypes.includes(this.data?.type)) {
+      // If the tile renderer is a tile server, the buildTileUrlTemplate function decides which assets it supports
+      else if (this.tileRendererType === 'server') {
         return true;
       }
       // Only http(s) links and relative links are supported
@@ -132,7 +134,7 @@ export default {
         return false;
       }
       // Otherwise, all images that a browser can read are supported + GeoJSON
-      else if (mapMediaTypes.includes(this.data?.type)) {
+      else if (mapTypes.includes(this.data?.type)) {
         return true;
       }
       return false;
@@ -238,6 +240,26 @@ export default {
     copyButtonText() {
       let where = (!this.isBrowserProtocol && this.from) ? 'withSource' : 'generic';
       return this.$t(`assets.copyUrl.${where}`, {source: this.from});
+    }
+  },
+  watch: {
+    data: {
+      immediate: true,
+      async handler(data) {
+        this.tileUrlTemplate = null;
+        if (!this.isAsset || !this.buildTileUrlTemplate) {
+          return;
+        }
+        try {
+          const url = await this.buildTileUrlTemplate(data);
+          // Ignore stale results if the asset changed in the meantime
+          if (this.data === data) {
+            this.tileUrlTemplate = url;
+          }
+        } catch (error) {
+          console.error(error);
+        }
+      }
     }
   },
   methods: {